人口構成データ取得部分をhookにまとめた
